Conflicts:
	addons/screensaver.xbmc.builtin.dim/resources/language/Amharic/strings.po
	addons/skin.confluence/media/flagging/video/bluray.png
	addons/skin.confluence/media/flagging/video/hdmv.png
	addons/skin.xeebo/media/flagging/video/bluray.png
	media/Splash.png
	xbmc/cores/AudioEngine/Utils/AEConvert.cpp

+++ xbmc/storage/cdioSupport.cpp
@@ -1,5 +1,5 @@
 /*
- *      Copyright (C) 2005-2012 Team XBMC
+ *      Copyright (C) 2005-2013 Team XBMC
  *      http://www.xbmc.org
  *
  *  This Program is free software; you can redistribute it and/or modify
@@ -926,7 +926,7 @@ UINT CCdIoSupport::MsfSeconds(msf_t *msf)
 //    the total length of the disk, and
 //    the number of tracks.
 
-ULONG CCdIoSupport::CddbDiscId()
+uint32_t CCdIoSupport::CddbDiscId()
 {
   CSingleLock lock(*m_cdio);
